Renamespace page provider
