+++ src/Factory/Action/AbstractActionTypeFactory.php
@@ -3,7 +3,7 @@
 namespace webignition\BasilParser\Factory\Action;
 
 use webignition\BasilParser\Model\Action\ActionInterface;
-use webignition\BasilParser\PageProvider\PageProviderInterface;
+use webignition\BasilParser\Provider\Page\PageProviderInterface;
 
 abstract class AbstractActionTypeFactory implements ActionTypeFactoryInterface
 {
