+++ src/Provider/Page/PageProviderInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace webignition\BasilParser\PageProvider;
+namespace webignition\BasilParser\Provider\Page;
 
 use webignition\BasilParser\Exception\NonRetrievablePageException;
 use webignition\BasilParser\Exception\UnknownPageException;
