+++ tests/Unit/Factory/AssertionFactoryTest.php
@@ -15,9 +15,9 @@ use webignition\BasilParser\Model\Value\ValueTypes;
 use webignition\BasilParser\Model\Identifier\Identifier;
 use webignition\BasilParser\Model\Identifier\IdentifierInterface;
 use webignition\BasilParser\Model\Identifier\IdentifierTypes;
-use webignition\BasilParser\PageProvider\EmptyPageProvider;
-use webignition\BasilParser\PageProvider\PageProviderInterface;
-use webignition\BasilParser\PageProvider\PopulatedPageProvider;
+use webignition\BasilParser\Provider\Page\EmptyPageProvider;
+use webignition\BasilParser\Provider\Page\PageProviderInterface;
+use webignition\BasilParser\Provider\Page\PopulatedPageProvider;
 use webignition\BasilParser\Tests\Services\AssertionFactoryFactory;
 
 class AssertionFactoryTest extends \PHPUnit\Framework\TestCase
